Starting this test
This is an Arduino code for injecting a little noise into a facebook feed and maybe making a few friends.  This is a simple code that can be uploaded on a USB Arduino-type microprocessor (like the "beetle") and plugged into a computer when you want to use it.  The facebook browser just needs to be open for it to start.  I apologize in advance if you're from facebook.  Wait no, sorry I don't...